fix: false discovery
This PR prevents the DHT claiming discovery of a peer that it may not have even discovered and also multiple discoveries of the same peer.

DHT was listening for 'peer-mux-established' from the switch and emitting a 'peer' event in response to it. Listening to 'peer-mux-established' implies that the peer has _already_ been discovered so there's no need to emit it again and moreover the peer may not have been discovered by the DHT in the first place!

License: MIT
Signed-off-by: Alan Shaw <alan.shaw@protocol.ai>