fix: do not throw on empty provider list
If no content providers can be found, just be an empty iterator
instead of throwing.

This is necessary because multiple findProv strategies are tried
higher up the stack, at which point we make the decision whether
to throw or not.

This makes the behaviour consistent with the delegated content
routing module.