chore: guard on table refresh running
We queue up ping requests to see if we should evict peers from the
routing table - if we are being shut down the requests will be
aborted so check to make sure we are still running before starting
another.

Also, don't evict a peer because the request failed if we aren't
running as it may have been cancelled which doesn't mean the remote
peer is invalid.