feat: remove dht.getPublicKey
The `dht.getPublicKey` method is a specialised `dht.get` so you
should just call that with the right key value.

This lets us remove the `libp2p-crypto` dependency as we were using
it to unmarshal the returned key - if you are interested in a
node's public key you're probably about to do something with it
in which case you probably have a dep on `libp2p-crypto` so no need
for this module to have one.

BREAKING CHANGE: `dht.getPublicKey` has been removed

chore: update iso random stream

chore: remove