fix: increase ping concurrency
After running an js-ipfs node for a week I noticed the lan routing table
ping queue had 200k items in it (😱).

Increasing the concurrency alleviates the queue pressure, so here we
increase the default concurrency to 10 and expose it as a config option,
as well as the timeout value to allow tuning by the user for their
specific application.