feat: switch to server mode automatically when addresses change
If the node is in client mode, and the `self:peer:update` event is
emitted, and the event detail contains publicly routable addresses,
switch to server mode.

If the event is emitted and we are in server mode and the event detail
contains only private addresses, switch back to client mode.