fix: use events to delay before self-query
Instead of debouncing and using timeouts to wait for DHT peers before
running the initial self-query, instead get the routing table to emit
events when peers are added or removed - if the table is empty when
we run the self-query, wait for the `peer:add` event before
continuing.

Improves startup time.